CAM: MBPP Dialog - Add gcode lines numeration
